refactor: Inline git package into hooks package, clean up error messages
